@@ -17,18 +17,19 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
-  yaml:
-  cupertino_icons: ^0.1.3
+  yaml: ^3.1.2
+  cupertino_icons: ^1.0.6
+  http: ^0.13.6
   flutter:
     sdk: flutter
-
+  global_configuration: ^2.0.0-nullsafety.1
 
 dev_dependencies:
-  provider:
-  test:
+  provider: ^6.1.2
+  test: ^1.22.0
   flutter_test:
     sdk: flutter
 
@@ -42,6 +43,9 @@ flutter:
   # included with your application, so that you can use the icons in
   # the material Icons class.
   uses-material-design: true
+
+  assets:
+    - assets/cfg/
 
   # To add assets to your application, add an assets section, like this:
   # assets:
